requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-session==0.5.0
flask-jwt-extended==4.6.0
python-dotenv==1.0.0
requests==2.31.0
waitress==2.1.2
bcrypt==4.1.2
cryptography==44.0.1
pyjwt==2.8.0
boto3==1.34.43
botocore==1.34.43
chromadb==0.4.18
openai==1.24.0
langchain==0.0.335
langchain-community==0.0.18
# If you get an error with langchain-community, try updating to a newer version:
# langchain-community==0.2.1
unstructured==0.10.30
unstructured-inference==0.7.21
pdf2image==1.16.3
pytesseract==0.3.10
PyPDF2==3.0.1
# poppler-utils (install separately with: apt-get install -y poppler-utils)
pydantic==2.5.2
huggingface-hub==0.19.4
torch==2.2.0
transformers==4.35.2
sentence-transformers==2.2.2
pillow==10.2.0
tqdm==4.66.1
markupsafe==2.1.3
werkzeug==2.3.7
jinja2==3.1.2
itsdangerous==2.1.2
onnxruntime==1.15.1
pyotp==2.9.0
email-validator==2.1.1
stripe==8.3.0
prometheus-client==0.20.0
fastapi==0.110.0
uvicorn==0.27.1 
gunicorn==21.2.0
